pyproject.toml
# Remove duplicate build-system section for setuptools
# [build-system]
# requires = ["setuptools>=64.0", "wheel"]
# build-backend = "setuptools.build_meta"

[project]
name = "moderntensor"
version = "0.1.0"
description = "Moderntensor CLI"
readme = "README.md"
# Use table format required by setuptools 68.0.0
license = { text = "MIT" }
# Attempt to explicitly override license file auto-detection
license-files = []
authors = [
  { name = "son", email = "sonlearn155@gmail.com" }
]
keywords = ["blockchain", "layer1", "proof-of-stake", "ai-validation", "zkml", "cryptocurrency"]
classifiers = [
  "Programming Language :: Python :: 3",
]
requires-python = ">=3.10"
dependencies = [
  # Hedera SDK (official)
  "hiero-sdk-python>=0.2.0",

  # Cryptography (for commit-reveal hashing, signatures)
  "cryptography>=42.0.0",

  # Web framework (dashboard API, REST endpoints)
  "fastapi>=0.115.0",
  "uvicorn>=0.34.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",

  # Networking
  "websockets>=13.0",
  "requests>=2.32.0",

  # AI Validator Agent — LLM providers
  "openai>=1.0.0",
  "anthropic>=0.30.0",
  "google-generativeai>=0.5.0",

  # Utilities
  "click>=8.1.0",
  "loguru>=0.7.0",
  "PyYAML>=6.0",
  "python-dotenv>=1.0.0",
  "colorama>=0.4.6",
  "python-multipart>=0.0.20",
  "typing_extensions>=4.12.0",
]

[project.urls]
homepage = "https://github.com/sonson0910"

# Scripts (entry_points) - replacement for 'console_scripts' in setup.py
[project.scripts]
mtcli = "sdk.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["sdk*"]

[project.optional-dependencies]
cli = ["click>=8.1.0", "rich>=13.0.0"]

[tool.pytest.ini_options]
markers = [
    "integration: mark a test as an integration test",
    "ignore:`network` argument will be deprecated in the future.:UserWarning"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Add Hatch build configuration to specify package directory
[tool.hatch.build.targets.wheel]
packages = ["sdk"]
